+++ .gitignore
@@ -32,3 +32,6 @@ ltmain.sh
 **/Makefile.in
 **/stamp-h1
 **/compile_commands.json
+
+*.toml
+!/conf_example.toml
