+++ .gitmodules
@@ -1,3 +1,6 @@
-[submodule "eRPC-arm"]
-	path = eRPC-arm
+[submodule "extern/tomlcpp"]
+	path = extern/tomlcpp
+	url = git@github.com:cktan/tomlcpp.git
+[submodule "extern/eRPC-arm"]
+	path = extern/eRPC-arm
 	url = git@github.com:Peter-JanGootzen/eRPC-arm.git
