+++ dpfs_fuse/Makefile.am
@@ -10,20 +10,17 @@ lib_LTLIBRARIES = libdpfs_fuse.la
 libdpfs_fuse_adir = $(includedir)/
 include_HEADERS = dpfs_fuse.h
 
-libdpfs_fuse_la_LIBADD = $(srcdir)/../dpfs_hal/libdpfs_hal.a \
-	$(srcdir)/../../src/libmlx_dev_emu.a $(srcdir)/../../src/libmlx_dev_emu_snap.a \
-	$(srcdir)/../eRPC-arm/build/liberpc.a \
-	$(IBVERBS_LDFLAGS) $(SNAP_LDFLAGS) $(PYTHON3_LDFLAGS) 
-
+libdpfs_fuse_la_LIBADD = $(srcdir)/../dpfs_hal/libdpfs_hal.la \
+	$(srcdir)/../extern/eRPC-arm/build/liberpc.a
+	 
 libdpfs_fuse_la_LDFLAGS = -lboost_system -lboost_thread \
-	--exclude-libs=libmlx_dev_emu.a --exclude-libs=libmlx_dev_emu_snap.a --exclude-libs=liberpc.a -fvisibility=hidden \
-	$(IBVERBS_LDFLAGS) $(SNAP_LDFLAGS) $(PYTHON3_LDFLAGS) 
+	--exclude-libs=liberpc.a -fvisibility=hidden
 
 libdpfs_fuse_la_CPPFLAGS  = $(BASE_CPPFLAGS) \
 	-I$(srcdir)/../../src $(SNAP_CFLAGS) \
-	-I$(srcdir)/../dpfs_hal \
-	-I$(srcdir)/../eRPC-arm/third_party/asio/include \
-	-I$(srcdir)/../eRPC-arm/src \
+	-I$(srcdir)/../dpfs_hal/include \
+	-I$(srcdir)/../extern/eRPC-arm/third_party/asio/include \
+	-I$(srcdir)/../extern/eRPC-arm/src \
 	-DERPC_INFINIBAND -Wno-address-of-packed-member # eRPC required flags for its headers
 
 libdpfs_fuse_la_SOURCES = dpfs_fuse.cpp
