--- dpfs_hal/cpu_latency.h
@@ -1,7 +1,0 @@
-#ifndef CPU_LATENCY_H
-#define CPU_LATENCY_H
-
-int start_low_latency(void);
-void stop_low_latency(void);
-
-#endif // CPU_LATENCY_H
