+++ dpfs_hal/src/cpu_latency.h
@@ -1,0 +1,14 @@
+/*
+#
+# Copyright 2023- IBM Inc. All rights reserved
+# SPDX-License-Identifier: LGPL-2.1-or-later
+#
+*/
+
+#ifndef CPU_LATENCY_H
+#define CPU_LATENCY_H
+
+int start_low_latency(void);
+void stop_low_latency(void);
+
+#endif // CPU_LATENCY_H
