+++ dpfs_nfs/main.c
@@ -1,53 +1,29 @@
 /*
 #
-# Copyright 2020- IBM Inc. All rights reserved
+# Copyright 2023- IBM Inc. All rights reserved
 # SPDX-License-Identifier: LGPL-2.1-or-later
 #
 */
 
 #include <getopt.h>
+#include <stdlib.h>
 #include "dpfs_fuse.h"
 #include "dpfs_nfs.h"
-#include <stdlib.h>
 
 void usage()
 {
-    printf("virtionfs [-p pf_id] [-v vf_id ] [-e emulation_manager_name] [-s server_ip] [-x export_path] [-d queue_depth] \n");
+    printf("virtionfs [-c config_path]\n");
 }
 
 int main(int argc, char **argv)
 {
-    int pf = -1;
-    int vf = -1;
-    char *emu_manager = NULL; // the rdma device name which supports being an emulation manager and virtio_fs emu
-    char *server = NULL;
-    char *export = NULL;
-    uint32_t nthreads = 1;
-    uint32_t queue_depth = 64;
+    char *conf = NULL;
 
     int opt;
-    while ((opt = getopt(argc, argv, "p:v:e:s:x:t:d:")) != -1) {
+    while ((opt = getopt(argc, argv, "c:")) != -1) {
         switch (opt) {
-            case 'p':
-                pf = atoi(optarg);
-                break;
-            case 'v':
-                vf = atoi(optarg);
-                break;
-            case 'e':
-                emu_manager = optarg;
-                break;
-            case 's':
-                server = optarg;
-                break;
-            case 'x':
-                export = optarg;
-                break;
-            case 't':
-                nthreads = strtoul(optarg, NULL, 10);
-                break;
-            case 'd':
-                queue_depth = strtoul(optarg, NULL, 10);
+            case 'c':
+                config = optarg;
                 break;
             default: /* '?' */
                 usage();
@@ -55,29 +31,6 @@ int main(int argc, char **argv)
         }
     }
 
-    struct virtiofs_emu_params emu_params;
-    // just for safety
-    memset(&emu_params, 0, sizeof(struct virtiofs_emu_params));
-
-    if (pf >= 0)
-        emu_params.pf_id = pf;
-    else {
-        fprintf(stderr, "You must supply a pf with -p\n");
-        usage();
-        exit(1);
-    }
-    if (vf >= 0)
-        emu_params.vf_id = vf;
-    else
-        emu_params.vf_id = -1;
-    
-    if (emu_manager != NULL) {
-        emu_params.emu_manager = emu_manager;
-    } else {
-        fprintf(stderr, "You must supply an emu manager name with -e\n");
-        usage();
-        exit(1);
-    }
     if (server == NULL) {
         fprintf(stderr, "You must supply a server IP with -s\n");
         usage();
@@ -88,7 +41,7 @@ int main(int argc, char **argv)
         usage();
         exit(1);
     }
-    printf("virtionfs starting up!\n");
+    printf("dpfs_nfs starting up!\n");
     printf("Connecting to %s:%s\n", server, export);
 
     emu_params.polling_interval_usec = 0;
@@ -96,7 +49,7 @@ int main(int argc, char **argv)
     emu_params.queue_depth = queue_depth;
     emu_params.tag = "virtionfs";
 
-    dpfs_nfs_main(server, export, false, false, nthreads, &emu_params);
+    dpfs_nfs_main(server, export, false, false, nthreads, conf);
 
     return 0;
 }
