--- dpfs_rpc_remote/main.cpp
@@ -1,15 +1,0 @@
-/*
-#
-# Copyright 2020- IBM Inc. All rights reserved
-# SPDX-License-Identifier: LGPL-2.1-or-later
-#
-*/
-
-int main(int argc, char **argv) {
-  // init eRPC
-  // allocate some eRPC buffers
-  // plug eRPC into dpfs_fuse
-  // busy while loop { eRPC event loop once }
-
-
-}
